move extended vite settings to root
